feat: 上传文件

+++ dto/message.go
@@ -30,6 +30,6 @@ type SendFileMessageRequest struct {
 	Filename    string `form:"filename" json:"filename" binding:"required"`
 	FileHash    string `form:"file_hash" json:"file_hash" binding:"required"`
 	FileSize    int64  `form:"file_size" json:"file_size" binding:"required"`
-	ChunkIndex  int    `form:"chunk_index" json:"chunk_index"`
-	TotalChunks int    `form:"total_chunks" json:"total_chunks" binding:"required"`
+	ChunkIndex  int64  `form:"chunk_index" json:"chunk_index"`
+	TotalChunks int64  `form:"total_chunks" json:"total_chunks" binding:"required"`
 }
